refactor : Car의 name 변수를 객체로 만들기
