+++ src/main/java/carracingwinner/domain/Car.java
@@ -2,10 +2,9 @@ package carracingwinner.domain;
 
 public class Car {
 
-    private static final int MAX_NAME_LENGTH = 5;
     private static final int MOVE_BOUND = 4;
 
-    private String name;
+    private Name name;
     private Position position;
 
     public Car(String name) {
@@ -13,15 +12,8 @@ public class Car {
     }
 
     public Car(String name, Position position) {
-        validateName(name);
-        this.name = name;
+        this.name = new Name(name);
         this.position = position;
-    }
-
-    private void validateName(String name) {
-        if (name.length() > MAX_NAME_LENGTH) {
-            throw new IllegalArgumentException("이름은 " + MAX_NAME_LENGTH + "자를 초과할 수 없습니다.");
-        }
     }
 
     public void move(int moveValue) {
@@ -31,11 +23,11 @@ public class Car {
     }
 
     public String showCurrentState() {
-        return name + " : " + position.getStringValue();
+        return name.getName() + " : " + position.getStringValue();
     }
 
     public String getName() {
-        return name;
+        return name.getName();
     }
 
     public Position getPosition() {
