+++ src/main/java/carracingwinner/domain/Name.java
@@ -1,0 +1,24 @@
+package carracingwinner.domain;
+
+public class Name {
+
+    private static final int MAX_NAME_LENGTH = 5;
+
+    private final String name;
+
+    public Name(String name) {
+        validateName(name);
+        this.name = name;
+    }
+
+    private void validateName(String name) {
+        if (name.length() > MAX_NAME_LENGTH) {
+            throw new IllegalArgumentException("이름은 " + MAX_NAME_LENGTH + "자를 초과할 수 없습니다.");
+        }
+    }
+
+    public String getName() {
+        return name;
+    }
+
+}
